大改版暫存

+++ backend/src/app/users/serializers.py
@@ -5,6 +5,11 @@ from rest_framework import serializers
 from .models import User, CommonParticipant
 
 
+def validate_ntub_email(email: str):
+    if not email.endswith('@ntub.edu.tw'):
+        raise serializers.ValidationError('This is not a ntub email.')
+
+
 class LoginSerializer(serializers.Serializer):
     access_token = serializers.CharField(help_text='Google OAuth2 AccessToken')
 
@@ -24,11 +29,16 @@ class UserSerializer(serializers.ModelSerializer):
 class CommonParticipantSerializer(serializers.ModelSerializer):
     creator = serializers.SerializerMethodField()
     participant = serializers.SerializerMethodField()
+    emails = serializers.ListField(
+        child=serializers.EmailField(validators=[validate_ntub_email]),
+        write_only=True,
+        required=True,
+    )
 
     class Meta:
         model = CommonParticipant
-        fields = ('id', 'title', 'creator', 'participant')
-        read_only_fields = ('id', 'creator')
+        fields = ('id', 'title', 'creator', 'participant', 'emails')
+        read_only_fields = ('id', 'creator', 'participant')
 
     def get_creator(self, common):
         return common.creator.email
@@ -36,19 +46,13 @@ class CommonParticipantSerializer(serializers.ModelSerializer):
     def get_participant(self, common):
         return common.participant.values_list('email', flat=True)
 
+    # def validate_emails(self, value):
+    #     for i, v in enumerate(value):
+    #         if not v.endswith('@ntub.edu.tw'):
+    #             raise serializers.ValidationError({i: v})
+    #     return value
 
-class CreateCommonParticipantSerializer(CommonParticipantSerializer):
-    emails = serializers.ListField(
-        child=serializers.EmailField(),
-        write_only=True,
-        required=True,
-    )
-
-    class Meta(CommonParticipantSerializer.Meta):
-        fields = ('id', 'title', 'creator', 'participant', 'emails')
-        read_only_fields = ('id', 'creator', 'participant')
-
-    def create_participant_for_common(self, common, emails):
+    def create_participant_from_common(self, common, emails):
         for email in emails:
             username, domin = email.split('@')
             if domin == 'ntub.edu.tw':
@@ -63,16 +67,24 @@ class CreateCommonParticipantSerializer(CommonParticipantSerializer):
         emails = validated_data.pop('emails')
         common_participant = CommonParticipant.objects.create(**validated_data)
         self.create_participant_for_common(common_participant, emails)
+
         return common_participant
 
-
-class UpdateCommonParticipantSerializer(CreateCommonParticipantSerializer):
     def update(self, instance, validated_data):
-        emails = validated_data.pop('emails')
+        emails = validated_data.pop('emails', None)
+        common_participant = super().update(instance, validated_data)
+
+        if emails is None:
+            return common_participant
+
         CommonParticipant.participant.through \
-            .objects.filter(commonparticipant=instance) \
+            .objects \
+            .filter(commonparticipant=instance) \
+            .exclude(user__email__in=emails) \
             .delete()
 
-        common_participant = super().update(instance, validated_data)
-        self.create_participant_for_common(common_participant, emails)
+        new_emails = set(emails) - set(User.objects.filter(email__in=emails).values_list('email', flat=True))
+        User.objects.bulk_create([User(username=email.split('@')[0], email=email) for email in new_emails])
+        common_participant.participant.add(*User.objects.filter(email__in=emails))
+
         return common_participant
